Externalize KVM Agent's option to change migration thread timeout
We noticed that sometimes migration in KVM fails due to a latency to move the VM process (domain) from the source host to the target.
This has been noticed in Rhel7 and CentOS7. An exception can be observed due to the default (hard-coded) timeout that is extrapolated and ACS can't find the migrated VM's domain in the target host.

This PR intends to externalize that option in the KVM agent configuration. Default timeout will remain 10 seconds.